CLOUDSTACK-9406: Enable IPv6 on interface for hypervisor communication
In the SSVM there is a single interface which is connected to the 'cloud0'
bridge on the hypervisor when running on KVM.

This patch enables IPv6 on that interface and lets the SSH daemon listen on
that address.

It results in the 'ip' output showing:

  2: eth0: <BROADCAST,MULTICAST,UP,LOWER_UP> mtu 1500 qdisc pfifo_fast state UP qlen 1000
      link/ether 0e:00:a9:fe:00:62 brd ff:ff:ff:ff:ff:ff
      inet 169.254.0.98/16 brd 169.254.255.255 scope global eth0
      inet6 fe80::c00:a9ff:fefe:62/64 scope link
         valid_lft forever preferred_lft forever

The sshd_config will contain these ListenAddress lines:

  ListenAddress 169.254.0.98
  ListenAddress [fe80::c00:a9ff:fefe:62%eth0]

From the KVM hypervisor you can now connect to this address with the 'cloudstack-ssh' command:

  $ cloudstack-ssh fe80::c00:a9ff:fefe:62%cloud0

After login the 'who' command will show a login over Link-Local:

  root@s-214-VM:~# who
  root     pts/0        2016-06-07 13:29 (fe80::fc56:5bff:fe56:b813%eth0)
  root@s-214-VM:~#

Currently SSH is the only service listening on Link-Local, but future applications might be
a HTTP REST API listening on this Link-Local Address removing the need for patching the
SSVM through a socket.

All communcation with the SSVM can be done over this Link-Local address/interface.

Signed-off-by: Wido den Hollander <wido@widodh.nl>