kvm: Use libvirt to clone RBD volume instead of rados-java
Libvirt has support to clone RBD volumes and handle this for us.

We no longer manually need to snapshot and clone a image, we can
just ask libvirt to do this for us.

It saves us code and also makes sure the new RBD images have all
the features enabled the source image has.

This requires at least libvirt 2.0.0 which is present in:

- Ubuntu 18.04
- CentOS 7.5

Signed-off-by: Wido den Hollander <wido@widodh.nl>