ipv6: Support IPv6 in Advanced Shared Networks
When using a Advanced Network with the GuestType shared we can support
IPv6 as traffic does not flow through the Virtual Router but is handled
by the router of the network.

When IPv6 with Router Advertisements is configured properly on this network
CloudStack can calculate the EUI-64 address a Instance will obtain using
SLAAC (RFC4862).

This is only done for networks which are 'Shared' and where a IPv6 CIDR and
Gateway has been configured.

In addition when creating a network IPv6 is better supported.

A CIDR provided to CloudStack should be exactly 64-bits in size to properly
support IPv6.

Security Grouping already supports IPv6 and therefor this check can be removed.

Signed-off-by: Wido den Hollander <wido@widodh.nl>