kvm: Remove dom0minmemory and rename variables
In the StartUpRoutingCommand we used to send back dom0MinMemory, but the
Management Server does not use this value in any of its calculations.

Reserved Memory is already subtracted from the total memory on the Host
by the Agent when reporting it's memory statistics to the Management Server.

The Management Server does not need to know how much Memory is reserved on
the Host.

Signed-off-by: Wido den Hollander <wido@widodh.nl>